Arregle los metodos http de pedido y pedido_detalle, ya se relacionan con user y rol, ademas hice cambios varios de normalizacion en el codigo

+++ backend/woofandbarf/src/main/java/org/proyecto_integrador/woofandbarf/controller/CartController.java
@@ -26,7 +26,7 @@ public class CartController {
 
     @PostMapping("/{cartId}/product/{productId}")
     public CartDetail addProduct(@PathVariable Integer cartId,
-                                 @PathVariable Long productId,
+                                 @PathVariable Integer productId,
                                  @RequestParam Integer quantity) {
         return cartService.addProductToCart(cartId, productId, quantity);
     }
