+++ backend/woofandbarf/src/main/java/org/proyecto_integrador/woofandbarf/exceptions/ProductNotFoundException.java
@@ -1,7 +1,7 @@
 package org.proyecto_integrador.woofandbarf.exceptions;
 
 public class ProductNotFoundException extends RuntimeException {
-    public ProductNotFoundException(Long id) {
+    public ProductNotFoundException(Integer id) {
         super("No se encontró el Product con id: " + id);
     }
 }
